+++ src/olmo_core/train/trainer.py
@@ -8,12 +8,14 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import (
     Any,
+    Callable,
     Dict,
     Generator,
     Literal,
     Optional,
     Tuple,
     TypedDict,
+    TypeVar,
     Union,
     cast,
 )
@@ -45,7 +47,7 @@ from ..nn.functional.cross_entropy_loss import (
     fused_cross_entropy_loss,
 )
 from ..optim import SkipStepOptimizer
-from ..utils import move_to_device
+from ..utils import cuda_sync_debug_mode, move_to_device
 from .callbacks import (
     Callback,
     CheckpointerCallback,
@@ -64,6 +66,8 @@ TRAIN_PPL_METRIC = "train/PPL"
 TRAIN_Z_LOSS_METRIC = "train/Z loss"
 OPTIM_STEP_SKIPPED_METRIC = "optim/step skipped"
 SEQ_LEN_METRIC = "data/sequence length"
+
+T = TypeVar("T")
 
 
 class TrainerStateDict(TypedDict):
@@ -254,6 +258,12 @@ class Trainer:
     not affect the learning rate schedule.
     """
 
+    async_bookkeeping: Optional[bool] = None
+    """
+    Do collective bookkeeping operations like reducing metrics asynchronously.
+    This requires a separate CPU-only backend, and will default to ``True`` if one is available.
+    """
+
     _metrics: Dict[int, Dict[str, torch.Tensor]] = field(default_factory=OrderedDict)
     _metrics_reduce_type: Dict[str, Optional[ReduceType]] = field(default_factory=dict)
     _canceled: bool = False
@@ -314,17 +324,18 @@ class Trainer:
 
         # Maybe create separate process group for bookkeeping.
         if self._bookkeeping_pg is None and is_distributed():
-            if backend_supports_cpu():
-                log.info("Creating new process group for bookkeeping")
+            if self.async_bookkeeping is None:
+                self.async_bookkeeping = backend_supports_cpu()
+            if self.async_bookkeeping:
+                if not backend_supports_cpu():
+                    raise OLMoConfigurationError(
+                        "A CPU-only backend is required for async bookkeeping"
+                    )
+                log.info("Creating new process group for async bookkeeping")
                 self._bookkeeping_pg = dist.new_group(
                     ranks=None
                     if self.dp_process_group is None
                     else dist.get_process_group_ranks(self.dp_process_group)
-                )
-            else:
-                log.warning(
-                    "No CPU backend configured, bookkeeping collectives will occur on the default "
-                    "backend and will be blocking. This may result in slower training throughput."
                 )
 
         # Check data loader configuration.
@@ -393,7 +404,7 @@ class Trainer:
             and self.global_step > 0
             and self.global_step % self.cancel_check_interval == 0
         ):
-            self.thread_pool.submit(self._check_if_canceled)
+            self.check_if_canceled()
 
         if self.is_canceled:
             return True
@@ -472,7 +483,7 @@ class Trainer:
         The device used for collective bookkeeping (non-training) operations that can potentially.
         use a different backend.
         """
-        if backend_supports_cpu():
+        if self.async_bookkeeping and backend_supports_cpu():
             return torch.device("cpu")
         else:
             return self.device
@@ -519,7 +530,7 @@ class Trainer:
         Asynchronously check if the run is canceled. Use :data:`is_canceled` to see the result.
         This needs to be called by all ranks at the same point in the training loop.
         """
-        self.thread_pool.submit(self._check_if_canceled)
+        self._run_bookkeeping_op(self._check_if_canceled)
 
     def fit(self):
         """
@@ -981,21 +992,51 @@ class Trainer:
         log.warning(msg)
         self.cancel_run(msg)
 
+    def _run_bookkeeping_op(
+        self, op: Callable[..., T], *args, cb: Optional[Callable[[T], None]] = None, **kwargs
+    ):
+        if (
+            self.async_bookkeeping
+            and self.bookkeeping_device.type == "cpu"
+            and self.bookkeeping_pg is not None
+        ):
+            # Can safely run in the thread pool.
+            future = self.thread_pool.submit(op, *args, **kwargs)
+            if cb is not None:
+
+                def callback(fut: Future[T]):
+                    try:
+                        cb(fut.result())  # type: ignore[misc]
+                    except BaseException as e:
+                        log.exception(e)
+                        self._error = e
+
+                future.add_done_callback(callback)
+        else:
+            result = op(*args, **kwargs)
+            if cb is not None:
+                cb(result)
+
     def _check_if_canceled(self):
         if self._canceled:
             return
 
         canceling_rank = self._canceling_rank if self._canceling_rank is not None else -1
-        canceling_rank = all_reduce_value(
-            canceling_rank, self.bookkeeping_device, op=dist.ReduceOp.MAX, group=self.bookkeeping_pg
-        )
-        if canceling_rank >= 0:
-            cancel_reason = scatter_object(self._cancel_reason, src=canceling_rank)
-            assert cancel_reason is not None
-            self._canceled = True
-            self._canceling_rank = canceling_rank
-            self._cancel_reason = cancel_reason
-            log.warning(f"Run canceled from rank {canceling_rank}. Reason: {cancel_reason}")
+        # NOTE: this is a known host-device sync (potentially) so we don't need the warning
+        with cuda_sync_debug_mode(0):
+            canceling_rank = all_reduce_value(
+                canceling_rank,
+                self.bookkeeping_device,
+                op=dist.ReduceOp.MAX,
+                group=self.bookkeeping_pg,
+            )
+            if canceling_rank >= 0:
+                cancel_reason = scatter_object(self._cancel_reason, src=canceling_rank)
+                assert cancel_reason is not None
+                self._canceled = True
+                self._canceling_rank = canceling_rank
+                self._cancel_reason = cancel_reason
+                log.warning(f"Run canceled from rank {canceling_rank}. Reason: {cancel_reason}")
 
     def _log_metrics(self):
         if not self._metrics:
@@ -1008,37 +1049,14 @@ class Trainer:
         # so CUDA training can continue.
         metrics_to_reduce = move_metrics(self._metrics, self.bookkeeping_device)
         self._metrics.clear()
-
-        if self.bookkeeping_device.type == "cpu" and self.bookkeeping_pg is not None:
-            # If we have a separate CPU backend and process group we can safely reduce
-            # metrics on CPU in a thread.
-            future = self.thread_pool.submit(
-                reduce_metrics,
-                metrics_to_reduce,
-                self._metrics_reduce_type,
-                self.bookkeeping_device,
-                process_group=self.bookkeeping_pg,
-            )
-
-            def callback(fut):
-                try:
-                    self._check_and_pass_on_metrics(fut.result())
-                except BaseException as e:
-                    log.exception(e)
-                    self._error = e
-
-            future.add_done_callback(callback)
-        else:
-            # Otherwise we have to reduce them now in the main thread.
-            # NOTE: if we're training on GPU and didn't have a host device sync above, this will
-            # trigger a host-device sync as we transfer the metrics back to CPU post-reducing.
-            metrics = reduce_metrics(
-                metrics_to_reduce,
-                self._metrics_reduce_type,
-                self.bookkeeping_device,
-                process_group=self.bookkeeping_pg,
-            )
-            self._check_and_pass_on_metrics(metrics)
+        self._run_bookkeeping_op(
+            reduce_metrics,
+            metrics_to_reduce,
+            self._metrics_reduce_type,
+            self.bookkeeping_device,
+            process_group=self.bookkeeping_pg,
+            cb=self._check_and_pass_on_metrics,
+        )
 
     def _check_and_pass_on_metrics(self, metrics: Dict[int, Dict[str, float]]):
         for step in sorted(metrics.keys()):