+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed a potential distributed deadlock bug when training without a separate CPU-only bookkeeping backend.
+- Removed some unnecessary host-device syncs in `olmo_core.distributed.utils`.
 
 ## [v1.6.0](https://github.com/allenai/OLMo-core/releases/tag/v1.6.0) - 2024-11-01