+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a potential distributed deadlock bug when training without a separate CPU-only bookkeeping backend.
 - Removed some unnecessary host-device syncs in `olmo_core.distributed.utils`.
+- Added `Trainer(Config).async_bookkeeping` field to toggle async bookkeeping.
 
 ## [v1.6.0](https://github.com/allenai/OLMo-core/releases/tag/v1.6.0) - 2024-11-01